Update to SDL 2
* Updated to SDL 2
* Updated to 64 bit compatibility
* Fixed piles of compiler warnings

+++ archi.cc
@@ -157,7 +157,7 @@ Uint16 file::getword(void) {
 }
 
 SDL_RWops *file::rwOps(void) {
-  return SDL_RWFromMem(buffer, fsize);
+  return SDL_RWFromMem(buffer, (int)fsize);
 }
